Splitting test class in two components
